refactor sandbox_app.py; move arg-parsing, config-building, and __main__ code to separate files
